Step 3a (drift fix): move covariate-shaped types from model/ to api/covariate/
Per DIR-PACKAGE-DRIFT-FIX-punit step 3a. Five types (and their
tests) move from org.javai.punit.model to org.javai.punit.api.covariate
where they sit alongside the existing covariate abstractions
(Covariate, CovariateProfile, CovariateAlignment, etc.):

- CovariateValue
- CovariateDeclaration
- DayGroupDefinition
- RegionGroupDefinition
- TimePeriodDefinition

Touched:
- git mv each main and test file from model/ to api/covariate/.
- Rewrite their `package` declarations to the new home.
- Add `import org.javai.punit.model.UseCaseAttributes` to
  CovariateDeclaration (it references UseCaseAttributes, which
  stays in model/ until step 3b moves it to api/).
- Update import paths in spec.model.ExecutionSpecification and
  spec.registry.SpecificationLoader (the only non-test consumers
  outside the moved set).

Architecture-test bookkeeping: noTopLevelModelPackage rule's
freeze-store shrunk from 16 → 10 entries (the 5 main classes and
their inner-class entries cleared). engineUnderInternal rule's
store unchanged (no movement into engine/).

Tests green across all four punit modules; punitexamples composite
build green.

Closes ArchUnit rule contribution: noTopLevelModelPackage (5 of
the remaining model classes removed; rule still has further entries
pending steps 3b and 3c).
