+++ punit-core/archunit_store/90083594-5684-45be-a3cf-87dc0d2f9509
@@ -1,8 +1,4 @@
 Class <org.javai.punit.model.BaselineProvenance> does reside in a package 'org.javai.punit.model..' in (BaselineProvenance.java:0)
-Class <org.javai.punit.model.CovariateDeclaration> does reside in a package 'org.javai.punit.model..' in (CovariateDeclaration.java:0)
-Class <org.javai.punit.model.CovariateValue$StringValue> does reside in a package 'org.javai.punit.model..' in (CovariateValue.java:0)
-Class <org.javai.punit.model.CovariateValue> does reside in a package 'org.javai.punit.model..' in (CovariateValue.java:0)
-Class <org.javai.punit.model.DayGroupDefinition> does reside in a package 'org.javai.punit.model..' in (DayGroupDefinition.java:0)
 Class <org.javai.punit.model.ExpirationPolicy> does reside in a package 'org.javai.punit.model..' in (ExpirationPolicy.java:0)
 Class <org.javai.punit.model.ExpirationStatus$Expired> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
 Class <org.javai.punit.model.ExpirationStatus$ExpiringImminently> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
@@ -10,7 +6,5 @@ Class <org.javai.punit.model.ExpirationStatus$ExpiringSoon> does reside in a pac
 Class <org.javai.punit.model.ExpirationStatus$NoExpiration> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
 Class <org.javai.punit.model.ExpirationStatus$Valid> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
 Class <org.javai.punit.model.ExpirationStatus> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.RegionGroupDefinition> does reside in a package 'org.javai.punit.model..' in (RegionGroupDefinition.java:0)
 Class <org.javai.punit.model.TerminationReason> does reside in a package 'org.javai.punit.model..' in (TerminationReason.java:0)
-Class <org.javai.punit.model.TimePeriodDefinition> does reside in a package 'org.javai.punit.model..' in (TimePeriodDefinition.java:0)
 Class <org.javai.punit.model.UseCaseAttributes> does reside in a package 'org.javai.punit.model..' in (UseCaseAttributes.java:0)
