+++ punit-core/src/main/java/org/javai/punit/api/covariate/CovariateValue.java
@@ -1,4 +1,4 @@
-package org.javai.punit.model;
+package org.javai.punit.api.covariate;
 
 import java.util.Objects;
