+++ punit-core/src/main/java/org/javai/punit/api/covariate/TimePeriodDefinition.java
@@ -1,4 +1,4 @@
-package org.javai.punit.model;
+package org.javai.punit.api.covariate;
 
 import java.time.LocalTime;
 import java.util.Objects;
