+++ punit-core/src/main/java/org/javai/punit/spec/model/ExecutionSpecification.java
@@ -6,8 +6,8 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import org.javai.punit.api.covariate.CovariateDeclaration;
 import org.javai.punit.api.covariate.CovariateProfile;
-import org.javai.punit.model.CovariateDeclaration;
 import org.javai.punit.model.ExpirationPolicy;
 import org.javai.punit.model.ExpirationStatus;
 import org.javai.punit.model.UseCaseAttributes;
